forces: add cubic interpolation scheme
This interpolation scheme is the cubic Hermite interpolation scheme for
two and three dimensions. It will be needed for the future tabulated
force module to interpolate the force and potential from a precalculated
lattice of potential energies (and its derivatives).

It currently supports two precision formats, single and double floating
point precision that should be defined by the user.