Add helper module to generate the coefficients for an interpolation scheme
In order to simplify the coefficient generation for the interpolation
schemes, tabulated_generator can generate the lattice with the
interpolation coefficients for a given potential.

Currently, only a host implementation is available. But since this
calculations needs to be done only once and coefficients can be stored
and loaded from a H5MD file, this should not be a bottleneck.