mdsim: add region module to select particles from a simulation domain
This module allows to select particles based on their position in the
simulation domain. Whether a particle is within a region or outside of
it, will be determined by another module, the geometry.

The GPU implementation of the selection is similar to the implementation
of the binning module, but there are only 2 bins: inside or outside the
region. After determining a "mask" (or bin) for each particle, the
particles are sorted by the bin and then the bin border is determined,
to calculate the number of particles in each bin.

The mask can also determined independently in order to enable operations
like intersections, unions, etc. of regions/particle groups.

Note that there is not yet a geometry implementation, this will follow
in a later commit.